refactor : Player 인터페이스 추상클래스로 변경
