+++ src/main/java/blackjack/domain/player/Dealer.java
@@ -5,37 +5,19 @@ import blackjack.domain.card.Cards;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Dealer implements Player {
+public class Dealer extends Player {
 
     public static final String DEALER_NAME = "딜러";
     private static final int DEALER_RECEIVE_STANDARD = 16;
     private static final int DEALER_OPEN_CARD_SIZE = 1;
 
-    private final String name = DEALER_NAME;
-    private final Cards cards;
-
     public Dealer() {
-        cards = new Cards();
-    }
-
-    @Override
-    public void receiveCard(final Card card) {
-        cards.save(card);
+        super(DEALER_NAME, new Cards());
     }
 
     @Override
     public List<Card> openCards() {
         return new ArrayList<>(cards.getCards().subList(0, DEALER_OPEN_CARD_SIZE));
-    }
-
-    @Override
-    public List<Card> showCards() {
-        return List.copyOf(cards.getCards());
-    }
-
-    @Override
-    public int calculateResult() {
-        return cards.calculateTotalPoint();
     }
 
     @Override
